rename: add package and move responseDTO and errorDTO
