+++ src/main/java/com/example/backend/controller/SocialingController.java
@@ -1,6 +1,6 @@
 package com.example.backend.controller;
 
-import com.example.backend.dto.ResponseDTO;
+import com.example.backend.dto.response.ResponseDTO;
 import com.example.backend.global.security.CustomUserDetails;
 import com.example.backend.service.user.UserService;
 import io.swagger.v3.oas.annotations.Operation;
