+++ src/main/java/com/example/backend/controller/UserController.java
@@ -2,8 +2,8 @@ package com.example.backend.controller;
 
 import com.example.backend.domain.User;
 import com.example.backend.domain.enumType.UserType;
-import com.example.backend.dto.*;
 import com.example.backend.dto.login.SnsJoinRequestDTO;
+import com.example.backend.dto.response.ResponseDTO;
 import com.example.backend.dto.user.UserDefaultJoinRequestDTO;
 import com.example.backend.dto.user.UserModifyRequestDTO;
 import com.example.backend.dto.user.UserProfileDTO;
