+++ src/main/java/com/example/backend/dto/response/ErrorDTO.java
@@ -1,4 +1,4 @@
-package com.example.backend.dto;
+package com.example.backend.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
