+++ src/main/java/com/example/backend/global/utils/ResponseUtils.java
@@ -1,6 +1,6 @@
 package com.example.backend.global.utils;
 
-import com.example.backend.dto.ResponseDTO;
+import com.example.backend.dto.response.ResponseDTO;
 import com.example.backend.dto.login.AuthDTO;
 import com.example.backend.global.security.AuthToken;
 import org.springframework.http.HttpHeaders;
